Add files via upload
Screen-Time-Analysis So this is how we can analyze the screen time of a user using the Python programming language. Screen Time Analysis is the 
task of analyzing and creating a report on which applications and websites are used by the user for how much time. I hope 
you liked this article on Screen Time Analysis using Python. Feel free to ask valuable questions in the comments section 
below.